Moving auxiliary scripts to their own module
Test to move the aux functions away from the L1 or L2 processing modules. Can re-assign location or name if needed.

Uploading my scratch notebook so I can use it in the office today
