+++ seastar/retrieval/__init__.py
@@ -1,1 +1,2 @@
 import seastar.retrieval.level2
+import seastar.retrieval.auxiliary
